Update of codebase:
Improve the calculation code to be implemented via pointers where
appropriate to marginally improve seed and efficiency of calculation.

As a result, some method signatures have changed to return status
codes/void values when implemented with pointers.

Add a simple CLI interface to the program to add clarity to usage, and
update the README.

The code now reads in from a given filename (or stdin) rather than the
stdin that was hardcoded prior.